Extracted interfaces from services: - Extracted interface from MovieService - Extracted interface from GenreService
