+++ src/main/java/com/khrushch/movieland/service/DefaultGenreService.java
@@ -1,0 +1,28 @@
+package com.khrushch.movieland.service;
+
+import com.khrushch.movieland.model.Genre;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class DefaultGenreService implements GenreService {
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    private GenreCacheService cacheService;
+
+    @Override
+    public List<Genre> getAll() {
+        List<Genre> genres = cacheService.getAll();
+        logger.debug("Fetched genres: {}", genres);
+        return genres;
+    }
+
+    @Autowired
+    public void setCacheService(GenreCacheService cacheService) {
+        this.cacheService = cacheService;
+    }
+}
