+++ src/main/java/com/khrushch/movieland/service/GenreService.java
@@ -1,27 +1,9 @@
 package com.khrushch.movieland.service;
 
 import com.khrushch.movieland.model.Genre;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-public class GenreService {
-    private final Logger logger = LoggerFactory.getLogger(getClass());
-
-    private GenreCacheService cacheService;
-
-    public List<Genre> getAll() {
-        List<Genre> genres = cacheService.getAll();
-        logger.debug("Fetched genres: {}", genres);
-        return genres;
-    }
-
-    @Autowired
-    public void setCacheService(GenreCacheService cacheService) {
-        this.cacheService = cacheService;
-    }
+public interface GenreService {
+    List<Genre> getAll();
 }
